added app directory tests
