+++ src/app/index.ts
@@ -1,3 +1,2 @@
 export { DataHandler } from './handlers';
-
 export { appResolvers } from './resolvers';
